requirements.txt
pinecone-client==2.2.1
weaviate-client==3.18.0
azure-search-documents==11.4.0
guardrails-ai==0.1.8
transformers==4.28.0
torch==2.0.0
sentence-transformers==2.2.2
ragas==0.0.7
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.0
pandas==2.0.1
scikit-learn==1.2.2